Add regisgterReducerFactory
